EDSL: replace existential constraint by GADT constructor
This introduces equality constraints on `s` and therefore embeds /
foregoes the need for an existential. Makes the code slightly simpler
and less constraint-heavy at the cost of requiring users use the `inv`
function to invert character classes.

Not a bad trade-off overall, considering it's also better for
readability / clarity and type ambiguity.